Refactored to requiring filename.es6
