+++ src/ActionCreator.es6.js
@@ -14,8 +14,8 @@
 
 'use strict';
 
-var dispatcher = require('./dispatcherInstance');
-var debug = require('./debug');
+var dispatcher = require('./dispatcherInstance.es6');
+var debug = require('./debug.es6');
 var each = require('../lib/each');
 var PropTypes = require('react/lib/ReactPropTypes');
 var invariant = require('invariant');
