+++ src/ConstantCollection.es6.js
