+++ test.js
@@ -1,0 +1,15 @@
+var FluxThis = require('./build/FluxThis');
+
+debugger;
+
+var store = new FluxThis.ImmutableStore({
+	init: function () {
+
+	},
+	public: {
+
+	},
+	private: {
+
+	}
+});
