+++ test/webpack.config.js
@@ -34,11 +34,11 @@ module.exports = {
         })
     ],
     resolve: {
-        extensions: ['', '.webpack.js', '.web.js', '.js', '.es6.js']
+        extensions: ['', '.webpack.js', '.web.js', '.js']
     },
     module: {
         loaders: [{
-             test: /\.js$/, exclude: /node_modules/, loader: 'babel-loader'
+             test: /\.es6\.js$/, exclude: /node_modules/, loader: 'babel-loader'
         }]
     }
 };
